+++ src/constants/commands.ts
@@ -1,3 +1,6 @@
 export const COMPARE_FILES = 'foldersCompare.compareFiles';
 export const CHOOSE_FOLDERS_AND_COMPARE = 'foldersCompare.chooseFoldersAndCompare';
+export const COMPARE_FOLDERS_AGAINST_WORKSPACE = 'foldersCompare.compareFoldersAgainstWorkspace';
+export const COMPARE_FOLDERS_AGAINST_EACH_OTHER = 'foldersCompare.compareFoldersAgainstEachOther';
+export const COMPARE_SELECTED_FOLDERS = 'foldersCompare.compareSelectedFolders';
 export const REFRESH = 'foldersCompare.refresh';