+++ src/providers/viewOnlyProvider.ts
@@ -1,4 +1,4 @@
-import { TreeDataProvider, TreeItem, EventEmitter, Event } from 'vscode';
+import { TreeDataProvider, TreeItem, EventEmitter, Event, Uri, TreeItemCollapsibleState } from 'vscode';
 import { File } from '../models/file';
 import { build } from '../services/tree-builder';
 
@@ -6,11 +6,11 @@ export class ViewOnlyProvider implements TreeDataProvider<File> {
   private _onDidChangeTreeData: EventEmitter<any | undefined> = new EventEmitter<any | undefined>();
   readonly onDidChangeTreeData: Event<any | undefined> = this._onDidChangeTreeData.event;
   private diffs: string[][] = [];
-  private workspaceRoot: string = '';
+  private rootPath: string = '';
 
-  update(diffs: string[][], workspaceRoot: string) {
+  update(diffs: string[][], rootPath: string) {
     this.diffs = diffs;
-    this.workspaceRoot = workspaceRoot;
+    this.rootPath = rootPath;
     this._onDidChangeTreeData.fire();
   }
 
@@ -22,9 +22,19 @@ export class ViewOnlyProvider implements TreeDataProvider<File> {
     if (element && element.children) {
       return element.children;
     }
+    const {treeItems} = build(this.diffs, this.rootPath);
     const children = [];
-    const tree = build(this.diffs, this.workspaceRoot);
-    children.push(...tree.treeItems);
+    if (this.rootPath) {
+      children.push(
+        new File(
+          this.rootPath,
+          'root',
+          TreeItemCollapsibleState.Expanded,
+          undefined,
+          treeItems
+        )
+      );
+    }
 
     return children;
 	}